new stuff

+++ src/main/java/com/wesdm/springhibernate/service/EmployeeService.java
@@ -15,7 +15,7 @@ public interface EmployeeService {
  
     Employee findBySsn(String ssn);
     
-    Employee getReference(int id);
+    Employee getReference(Long id);
     
     void updateEmployee(Employee employee);
